Wrap image in PhotoHero
Wrapped the [AspectRatio] widget in a [PhotoHero] widget. When tapped, it will push a named route with the format '/product/id'.

Note: the width is constrained by the width of the [Container] widgets in the [AsymmetricView] so it just needs to be larger enough to take up the entire constraint.