Change HomePage to handle async method
Since [ProductsRepostiory.loadProducts] was changed to an async function, [HomePage] needed to be changed to handle it accordingly. This was done by implementing a [FutureBuilder] widget, which shows a loading page while the [Future] is resolving.

Since the [Future] cannot be calculated in the [build] method (otherwise it will never resolve since it will be restarted during each build), the constructor is no longer constant.