Change [ProductPage] to handle async
Since [ProductsRepository.getById] was refactored as an async method, [ProductPage] had to be changed accordingly. This was done by implementing a [FutureBuilder] widget that displays a loading page while the [Future] is resolving.

Since the [Future] cannot be calculated from within the [build] method (since it will be restarted during each build and will never resolve), the constructor is no longer constant.

Many widgets in the clustered [build] method were extracted into private methods for readability.